initial commit:  intial playbook creation and genesis pull
This commit will begin the work to allow for a genesis deployment to have the
ability to post install the nvidia kernel modules and other features.  Right
now everything is pretty well hard coded, upcoming commits will address the
dynamic obtaining of the kernel module, and any additional drivers that may
be needed.