checks: Add code to support recipe checks
To account for recipes having multiple flavors we implemented a system to allow for a combination of common and unique scripts that will determine what pre-playbook steps need to be taken based on the passed in config file.

Every file in the scripts/deploychecks directory ending in *_check will be executed with the config file passed in.  The script will determine if it needs to take any actions, and if there are any errors in the config definition.